requirements.txt
asn1crypto==0.24.0
azure-cognitiveservices-speech==1.5.1
cffi==1.12.3
chardet==3.0.4
cryptography==2.7
idna==2.8
numpy==1.16.4
pycparser==2.19
pydub==0.23.1
pyOpenSSL==19.0.0
PySocks==1.7.0
requests==2.22.0
scipy==1.2.1
urllib3==1.24.2
websocket-client==0.56.0
websockets==7.0
win-inet-pton==1.1.0
